requirements.txt
numpy~=1.25.1
Pillow~=10.0.0
PySimpleGUI~=4.60.5
pytest~=7.4.0
